pyproject.toml
[project]
name = "devscope"
version = "0.1.0"
description = "Local dev environment inventory and management CLI"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Kent", email = "clementchiboko@gmail.com" }]
keywords = ["devtools", "docker", "postgres", "cli", "infrastructure"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Build Tools",
  "Topic :: System :: Systems Administration",
]
dependencies = [
  "typer>=0.12",
  "rich>=13.0",
  "textual>=0.60",
  "docker>=7.0",
  "psycopg2-binary>=2.9",
  "pydantic>=2.0",
  "psutil>=5.9",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "ruff>=0.4",
  "mypy>=1.10",
  "twine>=5.0",
]

[project.scripts]
devscope = "devscope.cli:app"

[project.urls]
Homepage = "https://github.com/kentchiboko/devscope"
Repository = "https://github.com/kentchiboko/devscope"
Issues = "https://github.com/kentchiboko/devscope/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/devscope"]

# ---------------------------------------------------------------------------
# uv
# ---------------------------------------------------------------------------
[tool.uv]
dev-dependencies = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "ruff>=0.4",
  "mypy>=1.10",
  "twine>=5.0",
]

# ---------------------------------------------------------------------------
# ruff
# ---------------------------------------------------------------------------
[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["devscope"]

# ---------------------------------------------------------------------------
# mypy
# ---------------------------------------------------------------------------
[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
show_error_codes = true
mypy_path = "src"

# ---------------------------------------------------------------------------
# pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=devscope --cov-report=term-missing"

[tool.coverage.run]
source = ["src/devscope"]
branch = true

[tool.coverage.report]
fail_under = 80
